pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "configdiff"
version = "0.1.0"
description = "Structure-aware configuration comparison CLI for semantic diff across JSON, YAML, TOML, and INI."
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Dani Issac", email = "daniissac654321@gmail.com" }]
keywords = ["config", "diff", "yaml", "json", "toml", "ini", "devops", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
configdiff = "configdiff.cli.app:main"

[project.urls]
Homepage = "https://github.com/daniissac/configdiff"
Repository = "https://github.com/daniissac/configdiff"
Issues = "https://github.com/daniissac/configdiff/issues"
Changelog = "https://github.com/daniissac/configdiff/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "hypothesis>=6.100",
    "ruff>=0.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.coverage.run]
source = ["configdiff"]

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 80

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.hatch.build.targets.wheel]
packages = ["configdiff"]

[tool.hatch.build.targets.sdist]
include = ["configdiff/", "LICENSE", "README.md", "CHANGELOG.md"]
